% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filr.R
\name{filr}
\alias{filr}
\title{Create a list of files used in your script}
\usage{
filr(...)
}
\arguments{
\item{...}{\code{pth} or \code{pth}-like objects}
}
\description{
Create a list of files used in your script
}
