% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filr.R
\name{pth}
\alias{pth}
\alias{as.pth}
\alias{as.pth.pth}
\alias{as.pth.character}
\title{Create a path object}
\usage{
pth(location, is_dep = TRUE)

as.pth(location, ...)

\method{as.pth}{pth}(location, ...)

\method{as.pth}{character}(location, ...)
}
\arguments{
\item{location}{A path to a file, relative to the project's root}

\item{is_dep}{Is the file a dependency of the current script?}
}
\description{
Create a path object
}
